Consolidate setret handling for diverging functions
Some functions never associate a mini IR source register with an LLVM
value because they unconditionally throw.
`System.Runtime.Intrinsics.X86.Aes:Decrypt` is an example of such a
function; when compiled for arm64, it never returns a value and
unconditionally throws a `PlatformNotSupportedException`.

We already had support for handling this for some, but not all, return
value passing conventions. Deduplicate this support and apply it
uniformly to all return value passing conventions that expect a
populated mini IR source register.